movie_recommender/requirements.txt
# requirements.txt
google-adk
chromadb
sentence-transformers
spacy
pandas
numpy
google-generativeai
torch
tqdm